refactor: implement feature-first hybrid architecture
- Created feature-first hybrid architecture for scalability
- Organized code into features/ (business logic) and shared/ (common code)
- Migrated auth feature completely (actions, lib, schemas)
- Migrated shared code (ui components, config, utils, types)
- Updated TypeScript path aliases (@/features/*, @/shared/*)
- Fixed all import paths across the application
- Updated CONTRIBUTING.md with architecture guidelines
- Created comprehensive documentation (README files, architecture plan)

Architecture Benefits:
- Better code organization and maintainability
- Clear separation of concerns
- Scalable structure for future EPICs
- Team-friendly for parallel development
- Reduced coupling between features

Directory Structure:
src/
├── app/         - Next.js routes only
├── features/    - Feature modules (self-contained business logic)
│   └── auth/    - Authentication feature
└── shared/      - Reusable code (components, config, utils)

See reports/FEATURE-FIRST-ARCHITECTURE-PLAN.md for details
