+++ reports/PR-DESCRIPTION.md
@@ -1,0 +1,53 @@
+## Admin User Management System
+
+**Closes #40**
+
+### Overview
+
+Implements complete user management system for administrators (US-5.3) with create, edit, deactivate functionality, audit logging, and security controls.
+
+### ✅ Acceptance Criteria - All Met
+
+- ✅ Admin can create new users with username, password, role
+- ✅ Admin can edit user details and roles
+- ✅ Admin can deactivate (not delete) users
+- ✅ Deactivated users cannot log in
+- ✅ User management actions are logged
+
+### Key Changes
+
+**Database:**
+- Added `is_active` column to users table
+- Created `user_management_logs` table for audit trail
+- Migration: `migrations/003_add_user_management.sql`
+
+**Backend:**
+- Server actions: create, update, deactivate, activate, list users
+- Well-modularized into 5 files (schemas, auth, audit, queries, actions)
+- Admin-only access with bcrypt password hashing
+- Input validation with Zod
+
+**Frontend:**
+- Enhanced admin dashboard with user management UI
+- User table with inline edit/deactivate actions
+- Create and edit user modals with form validation
+- Real-time statistics and recent activity log
+
+### Testing
+
+1. Run migration: `npm run migrate`
+2. Create test users: `node scripts/create-test-users.mjs`
+3. Login as `admin_test` / `admin123`
+4. Test create, edit, activate/deactivate features at `/admin`
+
+### Security & Quality
+
+✅ No hardcoded secrets | ✅ Parameterized SQL queries | ✅ Input validation  
+✅ Password hashing | ✅ Admin-only middleware | ✅ Audit logging  
+✅ TypeScript strict mode | ✅ Well-modularized code | ✅ No debug code
+
+**Full details:** See `US-5.3-IMPLEMENTATION-REPORT.md`
+
+---
+
+**Ready for review!** 🚀
