+++ reports/README.md
@@ -1,0 +1,28 @@
+# Implementation Reports
+
+This folder contains detailed implementation reports and documentation for completed user stories and issues.
+
+## Purpose
+
+- Keep implementation documentation separate from the main codebase
+- Provide detailed records of how features were implemented
+- Store PR descriptions and technical details for reference
+- Maintain a history of development work
+
+## Contents
+
+Each report typically includes:
+- User story/issue details
+- Acceptance criteria verification
+- Implementation approach
+- Files changed
+- Testing instructions
+- Security and compliance notes
+
+## Naming Convention
+
+Reports should follow the format: `US-X.Y-IMPLEMENTATION-REPORT.md` or `ISSUE-XX-REPORT.md`
+
+---
+
+**Note:** These reports are for documentation purposes and can be referenced in PRs but are not part of the production codebase.
