+++ src/app/api/health/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
-import { config } from '@/lib/config/env';
-import { performHealthCheck } from '@/lib/config/init';
-import { logger } from '@/lib/config/logger';
+import { config } from '@/shared/config/env';
+import { performHealthCheck } from '@/shared/config/init';
+import { logger } from '@/shared/config/logger';
 
 /**
  * GET /api/health
