+++ src/app/supervisor/page.tsx
@@ -1,7 +1,7 @@
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
+import { Card, CardContent, CardHeader, CardTitle } from "@/shared/components/ui/card"
 import { ClipboardList, AlertTriangle, Clock } from "lucide-react"
 
-import { verifySession } from "@/lib/session"
+import { verifySession } from "@/features/auth/lib/session"
 
 export default async function SupervisorDashboard() {
     const session = await verifySession()
