+++ src/shared/types/config.types.ts
